+++ src/ft_lstutils.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_lstutils.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mpuig-ma <mpuig-ma@student.42barcel>       +#+  +:+       +#+        */
+/*   By: jsebasti <jsebasti@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/05 16:12:37 by mpuig-ma          #+#    #+#             */
-/*   Updated: 2023/10/05 16:12:46 by mpuig-ma         ###   ########.fr       */
+/*   Updated: 2023/11/03 20:28:24 by jsebasti         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3D.h"
 
-size_t	ft_lstheight(t_list *list)
+int		ft_lstheight(t_list *list)
 {
-	size_t	len;
+	int		len;
 
 	len = 0;
 	while (list != NULL)
@@ -25,10 +25,10 @@ size_t	ft_lstheight(t_list *list)
 	return (len);
 }
 
-size_t	ft_lstwidth(t_list *list)
+int		ft_lstwidth(t_list *list)
 {
-	size_t	len;
-	size_t	line_len;
+	int		len;
+	int		line_len;
 	char	*line;
 
 	len = 0;
